server/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
bcrypt==4.0.1
boto3==1.35.63
botocore==1.35.63
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.111.1
fastapi-cli==0.0.5
fastapi-utils==0.8.0
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.27.2
idna==3.10
Jinja2==3.1.4
jmespath==1.0.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
motor==3.6.0
mypy-extensions==1.0.0
numpy==2.1.3
passlib==1.7.4
psutil==5.9.8
pydantic==2.8.2
pydantic-settings==2.6.1
pydantic_core==2.20.1
Pygments==2.18.0
PyJWT==2.3.0
pymongo==4.9.2
pypdf==5.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.17
pytz==2022.7.1
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
s3transfer==0.10.3
scipy==1.14.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sseclient==0.0.27
sseclient-py==1.8.0
starlette==0.37.2
typer==0.13.0
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.30.6
uvloop==0.21.0
watchfiles==0.24.0
websockets==14.1
